refactor(keystore): rm `KeyType`, `OwnedKeyType`
We no longer need to assume anything about serializabilitiy of keys!
