+++ keystore/src/traits/mod.rs
@@ -1,15 +1,15 @@
+mod dyn_entity_id;
 mod entity;
 mod entity_database_mutation;
 mod fetch_from_database;
-mod key_type;
 mod primary_key;
 mod searchable_entity;
 mod unique_entity;
 
+pub use dyn_entity_id::DynEntityId;
 pub use entity::{Entity, EntityGetBorrowed};
 pub use entity_database_mutation::{EntityDatabaseMutation, EntityDeleteBorrowed};
 pub use fetch_from_database::FetchFromDatabase;
-pub use key_type::{DynEntityId, KeyType, OwnedKeyType};
 pub use primary_key::{BorrowPrimaryKey, PrimaryKey};
 pub use searchable_entity::{DeletableBySearchKey, SearchableEntity};
 pub use unique_entity::{UniqueEntity, UniqueEntityExt, UniqueEntityImplementationHelper};
